@@ -1,6 +1,5 @@
 import itertools
 import warnings
-from packaging.version import Version
 
 import numpy as np
 import pandas as pd
@@ -221,13 +220,6 @@ class TestPointPlotting:
                 np.linspace(0, 0.0, 1.0, self.N), ax.collections[0].get_alpha()
             )
 
-    # TODO temporary skip for mpl 3.8.0.dev
-    # (https://github.com/matplotlib/matplotlib/issues/25162)
-    @pytest.mark.skipif(
-        Version(matplotlib.__version__) >= Version("3.8.0.dev")
-        and Version(matplotlib.__version__) < Version("3.8.0"),
-        reason="failing with matplotlib dev",
-    )
     def test_legend(self):
         with warnings.catch_warnings(record=True) as _:  # don't print warning
             # legend ignored if color is given.
@@ -458,6 +450,17 @@ class TestPointPlotting:
         df = self.df.copy()
         df["category"] = df["values"].astype("str")
         df.plot("category", missing_kwds={"facecolor": "none"}, legend=True)
+
+    def test_missing_aspect(self):
+        self.df.loc[0, "values"] = np.nan
+        ax = self.df.plot(
+            "values",
+            missing_kwds={"color": "r"},
+            categorical=True,
+            legend=True,
+            aspect=2,
+        )
+        assert ax.get_aspect() == 2
 
 
 class TestPointZPlotting:
@@ -1843,6 +1846,8 @@ def test_column_values():
     t2 = Polygon([(1, 0), (2, 0), (2, 1)])
     polys = GeoSeries([t1, t2], index=list("AB"))
     df = GeoDataFrame({"geometry": polys, "values": [0, 1]})
+    numeric_index_polys = GeoSeries([t1, t2], index=[0, 1])
+    numeric_index_df = GeoDataFrame({"geometry": numeric_index_polys, "values": [0, 1]})
 
     # Test with continuous values
     ax = df.plot(column="values")
@@ -1861,6 +1866,11 @@ def test_column_values():
     colors_series = ax.collections[0].get_facecolors()
     np.testing.assert_array_equal(colors, colors_series)
     ax = df.plot(column=df["values"].values, categorical=True)
+    colors_array = ax.collections[0].get_facecolors()
+    np.testing.assert_array_equal(colors, colors_array)
+
+    # Test with pd.Index
+    ax = numeric_index_df.plot(column=numeric_index_df.index, categorical=True)
     colors_array = ax.collections[0].get_facecolors()
     np.testing.assert_array_equal(colors, colors_array)
 
@@ -1924,9 +1934,9 @@ def _check_colors(N, actual_colors, expected_colors, alpha=None):
     ), "Different lengths of actual and expected colors!"
 
     for actual, expected in zip(all_actual_colors, expected_colors):
-        assert actual == conv.to_rgba(expected, alpha=alpha), "{} != {}".format(
-            actual, conv.to_rgba(expected, alpha=alpha)
-        )
+        assert actual == conv.to_rgba(
+            expected, alpha=alpha
+        ), f"{actual} != {conv.to_rgba(expected, alpha=alpha)}"
 
 
 def _style_to_linestring_onoffseq(linestyle, linewidth):
@@ -1955,7 +1965,7 @@ def _get_ax(fig, label):
     for ax in fig.axes:
         if ax.get_label() == label:
             return ax
-    raise ValueError("no ax found with label {0}".format(label))
+    raise ValueError(f"no ax found with label {label}")
 
 
 def _get_colorbar_ax(fig):
